service/APIforCSR/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.0.3
django-rest-framework==0.1.0
djangorestframework==3.11.0
numpy==1.19.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
tflite-runtime @ https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-win_amd64.whl
wincertstore==0.2
